refactor!: move internal lib in separate folder
