+++ src/library/components/assets/texture.ts
@@ -1,8 +1,6 @@
+// biome-ignore lint/complexity/noStaticOnlyClass: <explanation>
 export default class Texture {
-  constructor() {}
-
   static from = (label: string) => {
-
     const texture = localStorage.getItem(label);
 
     if (!texture) throw new Error(`Texture does not exist: ${label}`);
