+++ src/library/components/drawables/tilingSprite.ts
