+++ src/library/index.ts
@@ -1,0 +1,8 @@
+import Application from "./application";
+import Assets from "./components/assets/assets";
+import Texture from "./components/assets/texture";
+import Container from "./components/containers/container";
+import Text from "./components/drawables/text";
+import TilingSprite from "./components/drawables/tilingSprite";
+
+export { Application, Container, Assets, Texture, Text, TilingSprite };
